moved static functions to Extensions class per PR request
